Algumas funções foram movidas para global_utils.go
