+++ src/main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"crypto/rand"
 	"fmt"
 	"log"
 	"math/big"
@@ -35,6 +34,13 @@ type Ranges struct {
 }
 
 func main() {
+
+	// Define constantes de configuração:
+	const (
+		tickerTime2printStatistics = 5 * time.Second // Tempo para imprimir os dados processados no console
+		tickerTime2randomAddress   = 1 * time.Minute // Modo 3 - Aleatório - Ticker para obter um número aleatório dentro do range
+	)
+
 	green := color.New(color.FgGreen).SprintFunc()
 
 	exePath, err := os.Executable()
@@ -124,12 +130,12 @@ func main() {
 	}
 
 	// Ticker for periodic updates every 5 seconds
-	ticker := time.NewTicker(5 * time.Second)
+	ticker := time.NewTicker(tickerTime2printStatistics)
 	defer ticker.Stop()
 	done := make(chan struct{})
 
 	// Ticker para obter um número aleatório dentro do range a cada hora:
-	tickerRandom := time.NewTicker(30 * time.Minute)
+	tickerRandom := time.NewTicker(tickerTime2randomAddress)
 	defer tickerRandom.Stop()
 
 	// Variavel to update last processed wallet address
@@ -263,104 +269,4 @@ func worker(wallets *Wallets, privKeyChan <-chan *big.Int, resultChan chan<- *bi
 			}
 		}
 	}
-}
-
-// Essa função calcula a posição privKeyInt representa entre rangeMinInt e rangeMaxInt, em porcentagem
-func calculatePercentage(pos, min, max *big.Int) float64 {
-	// Calcula o intervalo total
-	totalRange := new(big.Int).Sub(max, min)
-
-	// Calcula a posição relativa de privKeyInt dentro do intervalo
-	relativePosition := new(big.Int).Sub(pos, min)
-
-	// Converte totalRange e relativePosition para float64
-	totalRangeFloat, _ := new(big.Float).SetInt(totalRange).Float64()
-	relativePositionFloat, _ := new(big.Float).SetInt(relativePosition).Float64()
-
-	// Calcula a porcentagem
-	percentage := (relativePositionFloat / totalRangeFloat) * 100
-
-	return percentage - 1
-}
-
-// Marton - A função recebe um tempo estimado para conclusão e retorna string com anos;
-func formatDuration(seconds float64) string {
-	// Converte os segundos em um valor inteiro
-	totalSeconds := int64(seconds)
-
-	// Define constantes para as durações
-	const (
-		secondsInMinute = 60
-		secondsInHour   = 60 * secondsInMinute
-		secondsInDay    = 24 * secondsInHour
-		secondsInMonth  = 30 * secondsInDay
-		secondsInYear   = 12 * secondsInMonth
-	)
-
-	years := totalSeconds / secondsInYear
-	totalSeconds %= secondsInYear
-
-	/*
-		months := totalSeconds / secondsInMonth
-		totalSeconds %= secondsInMonth
-
-		days := totalSeconds / secondsInDay
-		totalSeconds %= secondsInDay
-	*/
-
-	// Monta a string formatada
-	formattedDuration := fmt.Sprintf("%d anos" /*, %d mes(es), %d dia(s)"*/, years /*, months, days*/)
-	return formattedDuration
-}
-
-func updatePrivKeyInt(privKeyInt, rangeMinInt, rangeMaxInt *big.Int, modoSelecionado int) {
-	rangeSize := new(big.Int).Sub(rangeMaxInt, rangeMinInt)
-	randomNum, err := rand.Int(rand.Reader, rangeSize)
-	if err != nil {
-		fmt.Println("Erro ao gerar número aleatório:", err)
-		return
-	}
-	privKeyInt.Add(randomNum, rangeMinInt)
-
-	posicaoPercent := calculatePercentage(privKeyInt, rangeMinInt, rangeMaxInt) // Calcula a posição, em porcentagem
-	posicaoPercentStr := fmt.Sprintf("%.12f", posicaoPercent)                   // Formata com 12 casas decimais
-
-	// Representação gráfica da porcentagem:
-	graphicRepresentation := generateGraphicRepresentation(posicaoPercent)
-
-	fmt.Printf("\nMovendo para nova posição aleatória: %s ; Nova posição: %s  -  %s\n", privKeyInt.Text(16), posicaoPercentStr, graphicRepresentation)
-
-	if modoSelecionado == 3 {
-		if posicaoPercent < 4 {
-			fmt.Println("Não queremos processar os primeiros 4% do range. Vamos re-calcular nova posição aleatória.")
-			updatePrivKeyInt(privKeyInt, rangeMinInt, rangeMaxInt, modoSelecionado)
-		}
-		if posicaoPercent > 99 {
-			fmt.Println("Não queremos processar os últimos 1% do range. Vamos re-calcular nova posição aleatória.")
-			updatePrivKeyInt(privKeyInt, rangeMinInt, rangeMaxInt, modoSelecionado)
-		}
-	}
-
-	fmt.Println()
-
-}
-
-// Aqui eu gero uma representação gráfica, em ASCII, da porcentagem:
-func generateGraphicRepresentation(percentage float64) string {
-	totalLength := 50
-	position := int(percentage * float64(totalLength) / 100)
-
-	if position >= totalLength {
-		position = totalLength - 1
-	} else if position < 0 {
-		position = 0
-	}
-
-	representation := make([]rune, totalLength)
-	for i := range representation {
-		representation[i] = '_'
-	}
-	representation[position] = 'X'
-
-	return fmt.Sprintf("[%s]", string(representation))
 }
